Add `date` variable to front matter to pin `jekyll-feed` publish times.
Some RSS/Atom readers will treat an entry as changed even if it shares
the same ID if the date has changed. When Jekyll rebuilds the site,
without a `date` value in the front matter of the document, Jekyll uses
the current build time as the publication time for that document. This
means that even though ID values are not changing, publication times
are, which results in some feed readers showing duplicate entries for
posts in our event collection. This fixes that by ensuring that the
`date` is always the same as the event's `startDate`. This isn't
technically accurate, but gives us the desired behavior.