Do not publish Media in Resistance event until payment clears.
Despite a green light from Marisa Holmes, one of the conference
organizers, regarding booking payment for this event and having
submitted the invoice to Paper Tiger (papertiger.org), we received
from her that, and I quote:

"I really do want you all to come and value your labor. But I'm looking
at the budget for this and we could not probably pay everyone who is
doing workshops, and so I'd be making an exception which is unfair.
CopWatch will be there. Labor organizers. I really have to pay everyone
or no one."

Respectfully, I disagree. Conference organizers make agreements with
their participants. It is not an all or nothing scenario unless you
choose for that to be that the case.

Tech Learning Collective is not an organization that offers free labor.

Moreover, we have already publicized this event on behalf of the Media
in Resistence conference organizers. This means we have *already* done
free labor. This is an unacceptable way for Paper Tiger/Media in
Resistance to behave and as a result I am unpublishing the event
announcement from all promotional channels until payment is received in
full.